Implemented banned users list + reworked admin panel routing.

+++ src/store/features/users/slice.ts
@@ -59,20 +59,35 @@ export const createUser = createAsyncThunk('users/createUser', async (user: User
 
 type UpdateRequest = {
   user: UserRequest,
-  addClassrooms: (Classroom['id'])[],
-  removeClassrooms: (Classroom['id'])[]
+  addClassrooms?: (Classroom['id'])[],
+  removeClassrooms?: (Classroom['id'])[]
 };
 export const updateUser = createAsyncThunk('users/updateUser', async (request: UpdateRequest, thunkAPI) => {
   try {
-    if (request.addClassrooms.length > 0) {
+    if (request.addClassrooms && request.addClassrooms.length > 0) {
       await userService.addUserToClassrooms(request.user.id!, request.addClassrooms);
     }
 
-    if (request.removeClassrooms.length > 0) {
+    if (request.removeClassrooms && request.removeClassrooms.length > 0) {
       await userService.removeUserFromClassrooms(request.user.id!, request.removeClassrooms);
     }
 
     return await userService.updateUser(request.user);
+  }
+
+  catch (error: any) {
+    const message = error || 'error';
+    return thunkAPI.rejectWithValue(message);
+  }
+});
+
+type UpdateBanRequest = {
+  id: User['id'],
+  banned: User['banned']
+};
+export const updateBannedUser = createAsyncThunk('users/updateUser', async ({id, banned}: UpdateBanRequest, thunkAPI) => {
+  try {
+    return await userService.updateBannedUser(id, banned);
   }
 
   catch (error: any) {
